# Optimize stack,scale