--- a/augly/video/augmenters/ffmpeg/resolution.py
+++ b/augly/video/augmenters/ffmpeg/resolution.py
@@ -1,13 +1,12 @@
 #!/usr/bin/env python3
 # Copyright (c) Facebook, Inc. and its affiliates.
 
-from typing import Dict, Tuple
+from typing import List, Optional
 
-from augly.video.augmenters.ffmpeg.base_augmenter import BaseFFMPEGAugmenter
-from ffmpeg.nodes import FilterableStream
+from augly.video.augmenters.ffmpeg.base_augmenter import BaseVidgearFFMPEGAugmenter
 
 
-class VideoAugmenterByResolution(BaseFFMPEGAugmenter):
+class VideoAugmenterByResolution(BaseVidgearFFMPEGAugmenter):
     def __init__(self, resolution: float):
         assert (
             isinstance(resolution, (int, float)) and resolution > 0.0
@@ -15,20 +14,31 @@ class VideoAugmenterByResolution(BaseFFMPEGAugmenter):
 
         self.resolution = resolution
 
-    def add_augmenter(
-        self, in_stream: FilterableStream, **kwargs
-    ) -> Tuple[FilterableStream, Dict]:
+    def get_command(
+        self, video_path: str, output_path: Optional[str] = None, **kwargs
+    ) -> List[str]:
         """
         Alters the resolution of the video
 
-        @param in_stream: the FFMPEG object of the video
+        @param video_path: the path to the video to be augmented
 
-        @returns: a tuple containing the FFMPEG object with the augmentation
-            applied and a dictionary with any output arguments as necessary
+        @param output_path: the path in which the resulting video will be stored.
+            If not passed in, the original video file will be overwritten
+
+        @returns: a list of strings of the FFMPEG command if it were to be written
+            in a command line
         """
-        return (
-            in_stream.video.filter(
-                "scale", f"iw*{self.resolution}", f"ih*{self.resolution}"
-            ).filter("pad", **{"width": "ceil(iw/2)*2", "height": "ceil(ih/2)*2"}),
-            {},
-        )
+        command = [
+            "-y",
+            "-i",
+            video_path,
+            "-vf",
+            f"scale=height:ih*{self.resolution}:width=iw*{self.resolution},"
+            + "pad=width=ceil(iw/2)*2:height=ceil(ih/2)*2",
+            "-c:a",
+            "copy",
+            "-preset",
+            "ultrafast",
+            output_path,
+        ]
+        return command