--- a/augly/video/augmenters/ffmpeg/stack.py
+++ b/augly/video/augmenters/ffmpeg/stack.py
@@ -35,8 +35,8 @@ class VideoAugmenterByStack(BaseVidgearFFMPEGAugmenter):
         @param output_path: the path in which the resulting video will be stored.
             If not passed in, the original video file will be overwritten
 
-        @returns: a list of strings of the FFMPEG command if it were to be written
-            in a command line
+        @returns: a list of strings containing the CLI FFMPEG command for
+            the augmentation
         """
         video_info = get_video_info(video_path)