--- a/augly/video/augmenters/ffmpeg/stack.py
+++ b/augly/video/augmenters/ffmpeg/stack.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python3
 # Copyright (c) Facebook, Inc. and its affiliates.
 
-from typing import List, Optional
+from typing import List
 
 from augly.utils import pathmgr
 from augly.video.augmenters.ffmpeg.base_augmenter import BaseVidgearFFMPEGAugmenter
@@ -24,9 +24,7 @@ class VideoAugmenterByStack(BaseVidgearFFMPEGAugmenter):
         self.use_second_audio = use_second_audio
         self.orientation = orientation
 
-    def get_command(
-        self, video_path: str, output_path: Optional[str] = None
-    ) -> List[str]:
+    def get_command(self, video_path: str, output_path: str) -> List[str]:
         """
         Stacks two videos together