--- a/test.py
+++ b/test.py
@@ -1,0 +1,3 @@
+import augly.video.functional as af
+
+af.hstack('/tmp/test-video-2.mp4','/tmp/test-video-3.mp4','/Users/admin/res.mp4',True)